Moved Game models into types/
